refactor(storage): modularize OAuth credential service into submodules
Split the monolithic `oauth.rs` file into focused submodules:

**New structure:**
- `oauth/audit.rs` - Audit logging for OAuth credential operations
  - Defines audit event constants and detail builders
  - Provides functions to log and write OAuth audit entries
- `oauth/microsoft.rs` - Microsoft Graph OAuth implementation
  - Token exchange and refresh operations
  - PKCE challenge/verifier generation
  - Authorization URL construction
  - Client secret encryption/decryption for stored credentials
  - Metadata parsing and validation helpers
- `oauth/mod.rs` - Public API and orchestration
  - Authorization flow start/callback endpoints
  - Flow state management and validation
  - Credential creation after successful authorization
  - Error types for callback failures
- `oauth/provider.rs` - Token provider for runtime credential access
  - `MicrosoftGraphCredentialTokenProvider` with token refresh
  - In-memory token cache with expiry tracking
  - Automatic credential rotation recovery
  - Reauth-required status handling
- `oauth/tests.rs` - Comprehensive test coverage
  - Token provider refresh scenarios
  - Rotation recovery logic
  - Audit detail formatting
  - Metadata encryption/decryption

**Benefits:**
- Improved maintainability with clear separation of concerns
- Easier navigation (audit, protocol, orchestration, runtime, tests)
- Better testability with isolated components
- Reduced cognitive load per file
